+++ apps/backend-starter/src/controllers/user-custom.controller.ts
@@ -1,7 +1,9 @@
 import { Controller, Get, Query } from '@nestjs/common';
 import { User } from '@prisma/client';
 
-import { UserController, UserFindManyQueryDto } from '../generated/user';
+import { UserController } from '@generated/nestjs-models-rest';
+import { UserFindManyQueryDto } from '@generated/rest-dtos';
+
 
 @Controller(['user'])
 export class UserCustomController extends UserController {
@@ -9,7 +11,7 @@ export class UserCustomController extends UserController {
   public async findMany(
     @Query() queryDto: UserFindManyQueryDto,
   ): Promise<User[]> {
-    console.log('Controller has been overrided');
+    console.info('Controller has been overrided');
     const formatedParams = this.userRestDtoService.formatFindManyDto(queryDto);
     return this.userService.findMany(formatedParams);
   }