+++ packages/nestjs-authentication/src/authentication/strategies/jwt.strategy.ts
@@ -2,11 +2,10 @@ import { Inject, Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-jwt';
 
-import { UserService } from '../../generated/user/common';
-import { USER_SERVICE } from '../../generated/user/common/user-model.constant';
 import { JwtTokenPayload } from '../dtos';
 import { StrategyOptionsService } from '../services';
 
+import { USER_SERVICE, UserService } from '@generated/nestjs-models-common';
 import { User } from '@prisma/client';
 
 @Injectable()